pyproject.toml
[project]
name = "nse-analysis"
version = "0.1.0"
description = "Daily NSE stock analytics and reporting pipeline"
readme = "README.md"
authors = [
    { name = "Kevin Kipkoech", email = "105924200+Kevin7744@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.2",
    "pandas<3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "python-dateutil<2.9",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests==2.31.0",
    "rich>=14.3.2",
    "structlog>=25.5.0",
    "supabase>=2.28.0",
    "typer>=0.24.0",
]

[project.scripts]
nse-analysis = "nse_analysis.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pandas-stubs>=3.0.0.260204",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
